% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorta.R
\name{rnorta}
\alias{rnorta}
\title{Simulating Random Vectors using the NORTA Method}
\usage{
rnorta(R = R, cor.matrix = cor.matrix, distr = distr,
  qparameters = NULL)
}
\arguments{
\item{R}{integer indicating the sample size.}

\item{cor.matrix}{matrix indicating the correlation matrix of the
multivariate normal distribution employed in the NORTA method.}

\item{distr}{character string vector of length \code{ncol(cor.matrix)}
naming the quantile functions of the desired marginal distributions.}

\item{qparameters}{list of \code{ncol(cor.matrix)} lists indicating the
parameter values of the quantile functions specified by \code{distr}.}
}
\value{
Returns \code{R} random vectors of size \code{ncol(cor.matrix)} with
marginal distributions specified by \code{distr} (and \code{qparameters}).
}
\description{
Utility function to simulate random vectors with predefined marginal
distributions via the NORTA method.
}
\details{
Checks are made to ensure that \code{cor.matrix} is a positive definite
correlation matrix. The positive definiteness of \code{cor.matrix} is
assessed via eigenvalues.

The \eqn{t}-th character string in \code{distr} indicates the quantile
function of the \eqn{t}-th marginal distribution. See
\code{\link{Distributions}} for the most common distributions. Quantile
functions supported by other R packages are allowed provided that these
packages have been uploaded first. However, note that no checks are made to
ensure that the character strings in \code{distr} correspond to valid names
of quantile functions.

If \code{qparameters = NULL} then the default parameter values for the
quantile functions specified by \code{distr} are used. Otherwise,
\code{qparameters} should be provided as a list of \code{ncol(cor.matrix)}
lists such that the \eqn{t}-th list contains the desired parameter values of
the \eqn{t}-th quantile function.
}
\examples{
## An example with standard logistic as marginal distribution.
set.seed(1)
R <- 1000
LatentCorrelation <- toeplitz(c(1, rep(0.8, 2)))
LatentCorrelation
CommonMarginals <- rep("qlogis", 3)
SimLogistic <- rnorta(R = R, cor.matrix = LatentCorrelation, distr = CommonMarginals)

## The following lines exemplify the NORTA method.
set.seed(1)
SimNormal <- rsmvnorm(R = R, cor.matrix = LatentCorrelation)
all(SimLogistic == qlogis(pnorm(SimNormal)))

## Change the marginal distributions to standard normal, standard
## logistic and standard extreme value distribution.
set.seed(1)
DiffMarginals <- c("qnorm", "qlogis", "qgumbel")
SimDiffMars <- rnorta(R = R, cor.matrix = LatentCorrelation, distr = DiffMarginals)
cor(SimDiffMars)
colMeans(SimDiffMars)
apply(SimDiffMars, 2, sd)

## Same as above but using parameter values other than the default ones.
set.seed(1)
qpars <- list(c(mean = 1, sd = 9), c(location = 2, scale = 1), c(loc = 3, 
    scale = 1))
SimDiffMars2 <- rnorta(R = R, cor.matrix = LatentCorrelation, distr = DiffMarginals, 
    qparameters = qpars)
cor(SimDiffMars2)
colMeans(SimDiffMars2)
apply(SimDiffMars2, 2, sd)
}
\references{
Cario, M. C. and Nelson, B. L. (1997) \emph{Modeling and
generating random vectors with arbitrary marginal distributions and
correlation matrix}. Technical Report, Department of Industrial Engineering
and Management Sciences, Northwestern University, Evanston, Illinois.

Li, S. T. and Hammond, J. L. (1975) Generation of pseudorandom numbers with
specified univariate distributions and correlation coefficients. \emph{IEEE
Transactions on Systems, Man and Cybernetics} \bold{5}, 557--561.

Touloumis, A. (2016) Simulating Correlated Binary and Multinomial Responses
under Marginal Model Specification: The SimCorMultRes Package. \emph{The R
Journal} \bold{8}, 79--91.
}
\author{
Anestis Touloumis
}
